requirements.txt
pytorch
torchvision
mpose
einops
torchsummary
onnx
onnxruntime
wandb